# feat(scripts-*): speed up all `isConverged`, `getAllPackageInfo` calls/usage by ~95%