--- a/scripts/monorepo/src/getAllPackageInfo.js
+++ b/scripts/monorepo/src/getAllPackageInfo.js
@@ -16,9 +16,10 @@ let cwdForPackageInfo;
 
 /**
  * @returns {typeof packageInfo}
+ * @param {(metadata:{project:import('@nx/devkit').ProjectConfiguration;packageJson:import('nx/src/utils/package-json').PackageJson})=>boolean} [predicate]
  */
-function getAllPackageInfo() {
-  if (packageInfo && cwdForPackageInfo === process.cwd()) {
+function getAllPackageInfo(predicate) {
+  if (!predicate && packageInfo && cwdForPackageInfo === process.cwd()) {
     return packageInfo;
   }
 
@@ -28,9 +29,18 @@ function getAllPackageInfo() {
   cwdForPackageInfo = process.cwd();
 
   for (const [projectName, projectConfig] of projects) {
+    const packageJson = JSON.parse(
+      fs.readFileSync(path.join(workspaceRoot, projectConfig.root, 'package.json'), 'utf-8'),
+    );
+
+    if (predicate && !predicate({ project: projectConfig, packageJson })) {
+      continue;
+    }
+
     packageInfo[projectName] = {
       packagePath: projectConfig.root,
-      packageJson: JSON.parse(fs.readFileSync(path.join(workspaceRoot, projectConfig.root, 'package.json'), 'utf-8')),
+      packageJson,
+      projectConfig,
     };
   }