--- a/scripts/monorepo/src/index.d.ts
+++ b/scripts/monorepo/src/index.d.ts
@@ -3,7 +3,7 @@ export { getDependencies } from './getDependencies';
 export { default as findGitRoot } from './findGitRoot';
 export { default as findRepoDeps } from './findRepoDeps';
 export { default as getAllPackageInfo } from './getAllPackageInfo';
-export { isConvergedPackage, shipsAMD } from './isConvergedPackage';
+export { isConvergedPackage } from './isConvergedPackage';
 export { getAffectedPackages } from './getAffectedPackages';
 export { getDefaultEnvironmentVars } from './getDefaultEnvironmentVars';
 export { getProjectMetadata, workspaceRoot, getUncommittedFiles, getUntrackedFiles } from './utils';