--- a/scripts/monorepo/src/index.d.ts
+++ b/scripts/monorepo/src/index.d.ts
@@ -6,7 +6,7 @@ export { default as getAllPackageInfo } from './getAllPackageInfo';
 export { isConvergedPackage } from './isConvergedPackage';
 export { getAffectedPackages } from './getAffectedPackages';
 export { getDefaultEnvironmentVars } from './getDefaultEnvironmentVars';
-export { getProjectMetadata, workspaceRoot, getUncommittedFiles, getUntrackedFiles } from './utils';
+export { workspaceRoot, getUncommittedFiles, getUntrackedFiles } from './utils';
 export * as eslintConstants from './eslint-constants';
 export { getNthCommit } from './getNthCommit';
 export { tree, flushTreeChanges } from './tree';