--- a/scripts/monorepo/src/utils.js
+++ b/scripts/monorepo/src/utils.js
@@ -1,30 +1,8 @@
 const { execSync } = require('child_process');
 
-const { workspaceRoot, readProjectConfiguration } = require('@nx/devkit');
-
-const { tree } = require('./tree');
+const { workspaceRoot } = require('@nx/devkit');
 
 const TEN_MEGABYTES = 1024 * 10000;
-
-/**
- * Gets nx project metadata
- * @param {string} projectName - package name
- * @param {ReturnType<typeof import('@nx/devkit').getProjects>=} allProjects - all workspace projects
- * @returns {import('@nx/devkit').ProjectConfiguration}
- */
-function getProjectMetadata(projectName, allProjects) {
-  if (allProjects) {
-    const project = allProjects.get(projectName);
-
-    if (!project) {
-      throw new Error(`Project ${projectName} not found in workspace`);
-    }
-
-    return project;
-  }
-
-  return readProjectConfiguration(tree, projectName);
-}
 
 /**
  *
@@ -57,5 +35,4 @@ function getUntrackedFiles() {
 
 exports.getUncommittedFiles = getUncommittedFiles;
 exports.getUntrackedFiles = getUntrackedFiles;
-exports.getProjectMetadata = getProjectMetadata;
 exports.workspaceRoot = workspaceRoot;