--- a/scripts/tasks/src/presets.ts
+++ b/scripts/tasks/src/presets.ts
@@ -1,22 +1,22 @@
 import fs from 'fs';
 import path from 'path';
 
-import { isConvergedPackage, shipsAMD } from '@fluentui/scripts-monorepo';
 import { addResolvePath, condition, option, parallel, series, task } from 'just-scripts';
 
 import { apiExtractor } from './api-extractor';
 import { JustArgs, getJustArgv } from './argv';
-import { babel, hasBabel } from './babel';
+import { babel } from './babel';
 import { clean } from './clean';
 import { copy, copyCompiled } from './copy';
 import { eslint } from './eslint';
 import { generateApi } from './generate-api';
 import { jest as jestTask, jestWatch } from './jest';
 import { lintImportTaskAll, lintImportTaskAmdOnly } from './lint-imports';
+import { getRawMetadata } from './metadata-utils';
 import { postprocessTask } from './postprocess';
 import { postprocessAmdTask } from './postprocess-amd';
 import { prettier } from './prettier';
-import { hasSass, sass } from './sass';
+import { sass } from './sass';
 import { buildStorybookTask, startStorybookTask } from './storybook';
 import { swc } from './swc';
 import { ts } from './ts';
@@ -49,6 +49,7 @@ export function preset() {
   basicPreset();
 
   const args = getJustArgv();
+  const metadata = getRawMetadata(process.cwd());
 
   // eslint-disable-next-line @typescript-eslint/no-empty-function
   task('no-op', () => {}).cached!();
@@ -57,7 +58,7 @@ export function preset() {
   task('copy-compiled', copyCompiled);
   task('jest', jestTask);
   task('jest-watch', jestWatch);
-  task('sass', sass());
+  task('sass', sass(metadata.hasSass()));
   task('ts:postprocess', postprocessTask());
   task('postprocess:amd', postprocessAmdTask);
   task('ts:commonjs', ts.commonjs);
@@ -84,7 +85,7 @@ export function preset() {
       return parallel(
         'ts:commonjs',
         'ts:esm',
-        condition('ts:amd', () => !!args.production && !isConvergedPackage()),
+        condition('ts:amd', () => !!args.production && !metadata.isConverged()),
       );
     }
 
@@ -100,7 +101,7 @@ export function preset() {
       'ts:compile',
       'copy-compiled',
       'ts:postprocess',
-      condition('babel:postprocess', () => hasBabel()),
+      condition('babel:postprocess', () => metadata.hasBabel()),
     );
   });
 
@@ -119,7 +120,7 @@ export function preset() {
     const moduleFlag = args.module;
     return series(
       'swc:esm',
-      condition('babel:postprocess', () => hasBabel()),
+      condition('babel:postprocess', () => metadata.hasBabel()),
       resolveModuleCompilation(moduleFlag),
     );
   });
@@ -139,8 +140,8 @@ export function preset() {
       'sass',
       'ts',
       'api-extractor',
-      condition('lint-imports:all', () => !isConvergedPackage() && shipsAMD()),
-      condition('lint-imports:amd', () => isConvergedPackage() && shipsAMD()),
+      condition('lint-imports:all', () => !metadata.isConverged() && metadata.shipsAMD()),
+      condition('lint-imports:amd', () => metadata.isConverged() && metadata.shipsAMD()),
     ),
   ).cached!();
 
@@ -148,8 +149,9 @@ export function preset() {
     return series(
       'clean',
       'copy',
-      condition('sass', () => hasSass()),
+      condition('sass', () => metadata.hasSass()),
       parallel('swc:compile', 'generate-api'),
+      'swc:compile',
     );
   }).cached!();
 
@@ -163,7 +165,7 @@ export function preset() {
     if (!moduleFlag) {
       return parallel(
         'swc:commonjs',
-        condition('swc:amd', () => !!args.production && !isConvergedPackage()),
+        condition('swc:amd', () => Boolean(args.production) && !metadata.isConverged()),
       );
     }