<!--
Thank you for submitting a pull request!

Please verify that:
* [ ] Code is up-to-date with the `master` branch
* [ ] Your changes are covered by tests (if possible)
* [ ] You've run `yarn change` locally


PR flow tips:
* [ ] Try to start with a Draft PR
* [ ] Once you're ready (ideally the pipeline is passing) promote your PR to Ready for Review. This step will auto-assign reviewers for your PR.
-->

## Previous Behavior

<!-- This is the behavior we have today -->

## New Behavior

Using `readProjectConfiguration` api for more than 1 call is extremely expensive. This was happening in various parts of our infra:
- isConvergedPackage function
- shipsAmd
- generateConfig ( beachball ) 
- others ( see file changes )

This PR refactores critical API's in our core infra to simplify logic and make  things significantly faster 🚅🚅🚅


| command | Before | After |
|--------|--------|--------|
| yarn workspace @fluentui/scripts-beachball jest  | 31s | 0.2s  / **𝚫 99% FASTER** |
| node -r ./scripts/ts-node/register scripts/executors/tag-react-components.ts  | 22s | 0.8s  / **𝚫 96% FASTER** |
| lage build --to react-migration-v8-v9  | 241.12s | 220.16s  / **𝚫 9% FASTER** |

**running `lage build` (whole monorepo):** 

> **NOTE:**
>
> - not that not every task calls following APIs 4 times
> - these calculations are for non parallel execution ( which doesn't happen on CI - that's why "actuall" speed up gain is "only" 9% )

| api | 1 call | just-scripts 1 call | `lage build` -  200 (count of monorepo projects) runs |
|--------|--------|--------|--------|
| old `isConvergedPackage()` |  380ms | 4 * 380ms = 1520ms | 304s |
| new `metadata.isConverged()`| 1ms  | 4 * 1ms = 4ms | 0.8s / **𝚫99.8% faster** |



## Related Issue(s)

<!-- Please link the issue being fixed so it gets closed when this is merged. -->

- Fixes #